magboost/requirements.txt
Django>=5.0.0
djangorestframework>=3.15.0
django-cors-headers>=4.0.0
Pillow>=10.0.0
pytz
